documentos encoded

+++ apps/expedientes/model/entity/gestorexpedientedb.class.php
@@ -46,7 +46,7 @@ class GestorExpedienteDB Extends core\ClaseGestor {
 		$oDbl = $this->getoDbl();
 	    
 	    $sQuery = "SELECT e.id_expediente, e.asunto, e.ponente, e.json_antecedentes, items.id, items.tipo
-                    FROM expedientes e, jsonb_to_recordset(e.json_antecedentes) as items(id smallint,tipo text) 
+                    FROM expedientes e, jsonb_to_recordset(e.json_antecedentes) as items(id integer,tipo text) 
                     WHERE items.id=$id AND items.tipo='$tipo' ";
 	    
 		if (($oDbl->query($sQuery)) === FALSE) {
@@ -85,7 +85,7 @@ class GestorExpedienteDB Extends core\ClaseGestor {
                 $Where_visto = "";
 		}
 	    $sQuery = "SELECT e.id_expediente, e.asunto, e.ponente, e.json_preparar, items.id, items.visto 
-                    FROM expedientes e, jsonb_to_recordset(e.json_preparar) as items(id smallint,visto smallint) 
+                    FROM expedientes e, jsonb_to_recordset(e.json_preparar) as items(id integer,visto smallint) 
                     WHERE items.id=$id_cargo $Where_visto";
 	    
 		if (($oDbl->query($sQuery)) === FALSE) {
